serial/8250: add support for Ingenic jz4780 UARTs
The UART block used in Ingenic SoCs is advertised as NS16550 compatible,
however it isn't quite. The only difference relevant to basic usage is
the presence of a 'UART module enable' bit in the FCR register which
must be set at all times in order for the UART block to function. This
patch introduces a new UART port type for Ingenic jz4780 UARTs which
simply sets that bit & provides appropriate flags.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>