mmc: jz4740: add support for the JZ4780
Add support for the JZ4780 MMC controller to the jz47xx_mmc driver. There
are a few minor differences from the 4740 to the 4780 that need to be
handled, but otherwise the controllers behave the same. The IREG and IMASK
registers are expanded to 32 bits. Additionally, some error conditions are
now reported in both STATUS and IREG. Writing IREG before reading STATUS
causes the bits in STATUS to be cleared, so STATUS must be read first to
ensure we see and report error conditions correctly.

Signed-off-by: Alex Smith <alex.smith@imgtec.com>
Signed-off-by: Paul Cercueil <paul@crapouillou.net>
(cherry picked from commit 10353abe311beb9681349d4f70fe149b9a76b92d)

Change-Id: I59243d38838129563af16be324d19daa66e720cb
Signed-off-by: Dragan Cecavac <dragan.cecavac@mips.com>