rtc: jz4740: add support for JZ4780
The JZ4780 requires a few small modifications. Before writing an RTC
register, the registers must be made writeable via the write enable
register. This register does not exist on the 4740, so only do this
for the 4780. Also, the current timeout for waiting for WRDY is not
always sufficient, so increase it.

Change-Id: I914bad4d28b227acb80e3032057231f105630db8
Signed-off-by: Alex Smith <alex.smith@imgtec.com>
Signed-off-by: Dragan Cecavac <dragan.cecavac@mips.com>