gpu: drm: jz4780: Add LCD and HDMI driver
Adds LCD and HDMI driver for the jz4780.

Note: Upstream has diverged. HDMI phy driver has moved out of
staging and has been renamed. DRM_BRIDGE support has been added
to the phy.

This patch moves the 3.16 work on the driver to 'make it work'
on the CI20 with 3.18.

There will be subtle changed in the encoder/connector side.
CRTC should be the same.

Note: The jz4780 LCDC likes the framebuffer to be explicitly unpacked.
This is especially necessary for the dma to work.

Hence the DefaultFbBpp must be set to 32. Otherwise XOrg defaults to
24 and the driver doesn't work.

Sample Xorg.conf for modesetting

Section "Device"
        Identifier      "Video Device"
        Driver          "modesetting"
EndSection

Section "Monitor"
        Identifier      "Monitor"
        Option          "PreferredMode" "1920x1080"
EndSection

Section "Screen"
        Identifier      "Screen"
        Monitor         "Monitor"
        Device          "Video Device"
        DefaultFbBpp    32
EndSection

Signed-off-by: Zubair Lutfullah Kakakhel <Zubair.Kakakhel@imgtec.com>

Conflicts:
	drivers/gpu/drm/Kconfig
	drivers/gpu/drm/Makefile