dma: jz4780: fix in dma_tx_status. Check if jzchan->desc is NULL
Detected while audio debugging. This was causing a panic.

Add checks to see if jzchan->desc is NULL or not before further checks.

Signed-off-by: Zubair Lutfullah Kakakhel <Zubair.Kakakhel@imgtec.com>

---
Changes since V2
correct indentation
removed check for vdesc as it is implied

Changes since V1
Combined the if checks together in 1 if instead of cascaded. JamesH and Chris
feedback.