HACK: Enable 32Khz clock in rtc core
This clock is required by the wifi chip on the CI20.

There is no simple way to tell the wifi driver about the clock via DT
as it is dynamically registered when the rtc driver is loaded.

And the wifi driver doesn't support it yet either.
There was a debate on lkml about it. But nothing yet.

http://thread.gmane.org/gmane.linux.drivers.devicetree/61215

brcmfmac-sdio just assumes the clock is enabled. Which was done using
platform data and helper wrapper drivers before DT.

Not intended for upstream.

Change-Id: Id142aba520dee6c2b798489431424cd8a64195ca
Signed-off-by: Zubair Lutfullah Kakakhel <Zubair.Kakakhel@imgtec.com>
Signed-off-by: Dragan Cecavac <dragan.cecavac@mips.com>