sound: soc: jz4740: Add DT support to jz4740-i2s driver
This patch adds device tree support for the jz4740 driver.

It is to prepare the driver for subsequent changes to allow it to
be used by the jz4780 based CI20 board.

The CI20 is DT only.

codec sysclk is set and enabled during dai_probe as we cannot rely
on the board file to do that.

Signed-off-by: Zubair Lutfullah Kakakhel <Zubair.Kakakhel@imgtec.com>

---
Upstreamable now