sound: soc: jz4740: Add jz4780 support to jz4740-i2s driver
The jz4780 i2s core is almost exactly the same as the jz4780.

The exception is different bit masks for TX/RX thresholds.

This patch adds a few checks and allows jz4780 based systems to use
the same driver.

Tested on the CI20

Signed-off-by: Zubair Lutfullah Kakakhel <Zubair.Kakakhel@imgtec.com>

---
DMA request types for the JZ4740 are passed statically in the driver
using the slave_config->slave_id variable.
However the JZ4780 DMA driver uses DT to handle that.