mips: ci20: Add BT regulator hacks to dt
These regulators need to be enabled for the chip to be ON.

Once it is ON, bluez/ttyS2 is only required. Broadcoms patchram
utility communicates with the chip and bluetooth works.

There is no simple driver fix. These can be in a misc driver.
Or like this and enabled in DT.