mips: jz4780: Use jz4780-tcu driver and DT instead of jz4740 timer driver
Eventually we'd like the tcu driver to be upstreamed and then jz4740
and jz4775 use that driver.

The alternate would be to add more code to arch/mips/jz4740/time.c which
isn't ideal.

Note:
jz4740-timer_init is still left called because the reset code depends on
the watchdog which depends on the timer base address variable being initialized.