MIPS: jz4780: Use jz4780-tcu driver and DT instead of jz4740 timer driver
Eventually we'd like the tcu driver to be upstreamed and then jz4740
and jz4775 use that driver.

The alternate would be to add more code to arch/mips/jz4740/time.c which
isn't ideal.

Note:
jz4740-timer_init is still left called because the reset code depends on
the watchdog which depends on the timer base address variable being initialized.

Change-Id: Ib8dea25e00ab713c8725a84466e9ba77717cd559
Signed-off-by: Dragan Cecavac <dragan.cecavac@mips.com>