usb: otg: jz4780_platform: Switch off OTG PHY overcurrent warnings
Previously, the kernel log would be spammed with messages about an
overcurrent change being detected on the OTG port. Ingenic worked
around this in their 3.0.8 kernel by setting a bit in one of the
OTG controller's registers that would disable the warning. This
commit implements the same workaround, but for the 3.18 kernel.