net: ethernet: dm9000: Accept a MAC address as a bootarg for CI20.
The MAC address for the CI20's DM9000 is stored in EEPROM on a different
part of the board. Previously, the DM9000 would make up a random MAC
after failing to read its own EEPROM. This commit makes it possible for
U-BOOT to pass the MAC address to the DM9000 driver. Such as:

bootargs="dm9000.mac_addr=${ethaddr}"

Note: This isn't the "correct" way to do this, but will do for now. The correct
method would be to recompile U-BOOT with CONFIG_OF_LIBFDT and CONFIG_FIT
enabled. Then the kernel should be compiled as a FIT image and U-BOOT can then
add an entry to the DM9000's DT node to set the MAC address (local-mac-address
= [xx xx xx xx xx xx];).

Change-Id: I25e77087122fb3ebceec20de224c1f2fc7624726
Signed-off-by: Harvey Hunt <harvey.hunt@imgtec.com>
Signed-off-by: Dragan Cecavac <dragan.cecavac@mips.com>