MIPS: JZ4780: Use generic GPIO definitions for JZ4780
The JZ4740-specific GPIO functions are not compiled when MACH_JZ4780 is
selected. This leads to a link error if gpio_to_irq is used when
compiling for JZ4780. Therefore, wrap all of the JZ4740-specific GPIO
definitions in #ifdef, and when compiling for JZ4780, instead #define
everything (including gpio_to_irq) to use the generic gpiolib
implementations.

Signed-off-by: Alex Smith <alex.smith@imgtec.com>