MIPS: jz4780: Copy loops_per_jiffy for core 0 to core 1
Commit 600e7a200136 sets a bit in Config7 which stops the XBurst core
from special-casing short loops to avoid branch target buffer lookups.
The default behaviour vastly slows down tight loops and thus results
in a low BogoMIPS/loops_per_jiffy value, which is used for the *delay()
functions. Setting this bit results in a higher BogoMIPS/loops_per_jiffy
value.

However, even though that bit also gets set on the second core, for
reasons I cannot figure out, it does not appear to take effect until
later on. The result is that when the delay calibration is run on the
second core it will calculate a low value for loops_per_jiffy (and at
that point delays using the calibrated value will delay for the correct
amount of time), but later on delays will delay for far too short a time.
This can be observed with udelay_test: using taskset to restrict
udelay_test.sh to core 0 results in all tests passing, on core 1 however
all tests fail.

Ingenic's kernel does not suffer from this problem yet I cannot see why,
they set the short loop BTB lookup flag in the same place we do, but
their kernel correctly calibrates the delay loop.

So, for now, as a workaround, copy the loops_per_jiffy value from core 0
to core 1. With this, both cores are able to pass udelay_test.

Signed-off-by: Alex Smith <alex.smith@imgtec.com>