MIPS: jz4740: Fix power off for JZ4780
The RTC on the JZ4780 requires a magic dance before each register write
to make the registers writeable, which is present in the RTC driver but
missing from the power off code. Without it the power off code will hang
waiting for the RTC to become ready.

Copy over this code from the RTC driver so that the board can be powered
off.

Signed-off-by: Alex Smith <alex.smith@imgtec.com>