MIPS: Handle KSEG1 addresses in mips_dma_mmap()
mips_dma_mmap() is passed a virtual address from dma_alloc_coherent()
which may be a KSEG1 address. In that case the conversion to a physical
address using virt_to_page() is incorrect. By chance this causes no
problem on the JZ4780 because the low 256MB of DRAM is aliased at
0x20000000, meaning the incorrect physical address actually points to
the same memory. Fix for the sake of correctness however.

Signed-off-by: Alex Smith <alex.smith@imgtec.com>