cpuidle: jz4780: Add Ingenic JZ4780 cpuidle driver
The JZ4780 has a high overhead to executing a MIPS wait on SMP, as a
core must flush out dirty cache lines from its data cache before doing
so. This is because the core clock is gated during a wait and if the
other core tries to access a dirty line from the waiting core's cache,
it will lock up.

To mitigate some of this impact, this driver provides a simple polling
top level idle state, to try to avoid the cache flushing overhead when
the wait will only be short. The second level state is implemented with
the MIPS wait instruction.

TODO: The exit latency and target residency values for the wait state
could probably be tweaked further to try to get more time in the wait
state without affecting performance, but it seems to be working pretty
well for now - almost no time spent in wait state under heavy load, but
most time spent in it while idle.

Signed-off-by: Alex Smith <alex.smith@imgtec.com>