UBIFS: TEMP: skip lazy cache flush for UBIFS
When flush_dcache_page in called in do_readpage in UBIFS
for executable mappings, the data cache is not actually flushed,
but the flush is postponed until flush_icache_page is called.

This approach seems to cause cache coherency issues that might be
sourced either to the cache flush handling for MIPS or by an
incorrect sequence of operations followed by UBIFS.

Until the source of the problem is pin-pointed, call __flush_dcache_page
in do_readpage directly which will result in an immediate data flush
for executable pages.

Change-Id: I384aa9149f375be76e5028ce762860426907b220
Signed-off-by: Ionela Voinescu <ionela.voinescu@imgtec.com>