ci20: Enable snps,dwc2 and USB2.0 jz4780 PHY driver
To follow trends on CI20 platform, we need to enable
USB driver which can be registered to the UDC framework
and support dual role USB configuration. This functionalities
are provided with the latest DWC2 driver which can be used
with the new USB2.0 phy driver (jz4780-usb-phy) which was
written specifically for the jz4780 SoC (CI20).

This patch enables DWC2 driver (snps,dwc2) host mode support
and new USB2.0 phy driver (jz4780-usb-phy).

Change-Id: I655f138bdfc938b704a77154fb2419d08e0d557f
Signed-off-by: Miodrag Dinic <miodrag.dinic@imgtec.com>