ci20: Enable snps,dwc2 driver and ConfigFS support
Latest Android development has moved to using ConfigFS
for various USB functions (ADB/MTP/PTP...) configurable
through the USB gadget framework.

To follow this trends on CI20 platform, we need to enable
USB driver which needs to be registered to the UDC framework
and support dual role USB configuration. This functionalities
are provided with the latest DWC2 driver which can be used
with the new USB2.0 phy driver (jz4780-usb-phy) which was
written specifically for the jz4780 SoC (CI20).

This patch enables all prerequisites for using the DWC2
driver (snps,dwc2) on the CI20 platform and enables
ConfigFS support for the following gadget functions:

[*]       Generic serial bulk in/out
[*]       Abstract Control Model (CDC ACM)
[*]       Mass storage
[*]       Function filesystem (FunctionFS)
[*]       MTP gadget
[*]         PTP gadget
[*]       Accessory gadget
[*]         Audio Source gadget
[*]       Uevent notification of Gadget state
[*]       MIDI function

This kernel support needs to be followed by appropriate
changes in userspace for CI20 platform, similarly to
what is described in the following article:

https://wiki.linaro.org/LMG/Kernel/AndroidConfigFSGadgets

Change-Id: I655f138bdfc938b704a77154fb2419d08e0d557f
Signed-off-by: Miodrag Dinic <miodrag.dinic@imgtec.com>