MIPS: zImage: Introduce cache_flush_all() interface
This commit introduces more elegant solution for cache flushing.
New interface implementation is simplified version of the
cache flushing mechanism used by the kernel :
    arch/mips/include/asm/r4kcache.h

The new approach fixes a bug in previous implementation of
flushcaches() which lead to a kernel hang when a compressed
kernel size crosses a threshold of around 6.7MB.

Currently cache_flush_all() implementation is limited only to
MACH_JZ4780 (CI20), but proposed design is more flexible to
extensions for other platforms (which is unlikely to happen,
but nevertheless).

If zImage is used for any other platform then CI20, an empty
cache_flush_all() implementation is used, but the warning is
printed out to inform the user.

Change-Id: I2a6dd24b794b771c0a3a9524ed17e27dea633d8a
Signed-off-by: Miodrag Dinic <miodrag.dinic@imgtec.com>