MIPS: zImage: Increase .text section offset
If a compressed kernel size crosses a threshold of around 6.7MB
zImage de-compressor starts overwriting the kernel .text section.
The kernel .text section offset is hardcoded and it was assumed
that the mentioned scenario wouldn't be triggered.

Since the current compressed kernel size is larger than 6.7MB
we need to raise .text section offset from 0x80F00000 to
0x81F00000 in order to avoid overwriting. This moves the .text
section upwards by 16MB which should be sufficient for a
reasonable amount of time.

It is suitable to have .text section offset hardcoded, because it
needs to be accompanied by the same value when creating Android boot.img,
which is also hardcoded in BoardConfig.mk in CI20 device project.

Change-Id: I3598026c0967189d8caf848892ad9273b4ef0d7f
Signed-off-by: Dragan Cecavac <dragan.cecavac@imgtec.com>