jz4780: setup: Reserve 0xf4000000-0xf4ffffff range used for TCSM
0xf4000000-0xf403fff area is used for TCSM according
to the data sheet. In practice we encounter lazy decoding
based on the upper 8 bits, which results that the whole
address range 0xf4XXX000-0xf4XXXfff gets mapped to TCSM.

This brings unpredictable behaviour and will likely cause
a kernel crash once this memory range gets allocated.

__get_vm_area is used to reserve this range.
PAGE_SIZE is subtracted from the total size in order
to provide space for the guard page.

Change-Id: Ibb3e18e2833542f1b240f109365fe2b8d1a38004
Signed-off-by: Dragan Cecavac <dragan.cecavac@imgtec.com>