UBI: android: Implement BLKROSET dummy ioctl
Enable usage of "adb remount" command.
Do not report error when BLKROSET ioctl is called for UBI device.

This ioctl call is used on Android due to a security breach
where filesystem could be mounted read-only, but a user
would be able to write raw bytes to block device and
make modifications to the stored data.

More info regarding the security breach:
https://android.googlesource.com/platform/system/core/+/e18c0d5%5E!/

Later changes which check the return value of this ioctl
introduced the regression for running "adb remount" on UBI.

Testing has been performed and it showed that UBI
devices can not be written to, so additional level
of protection like this one is not necessary.

Change-Id: I362262a529e67cecdb4122d489939d537c46a976
Signed-off-by: Dragan Cecavac <dragan.cecavac@imgtec.com>