jz4780_camera: Fix image offset error
Data sheet recommends reseting rx fifo and setting status to 0
before turning CIM on. This commit makes sure that these steps
are not omitted. All code related to enabling/disabling
is placed in inline functions, which are being called when needed.

Without this change camera driver may start filling
image buffer with a short, but noticeable offset, which is
actually the last part of the previous image which has been
left out in rx fifo buffer. The result is image whose
far right part is moved to the far left.

Change-Id: I670fa674c05a2685be038d606e3d41503f217ed4
Signed-off-by: Dragan Cecavac <dragan.cecavac@imgtec.com>