UBIFS: Implement fallocate FALLOC_FL_PUNCH_HOLE|FALLOC_FL_KEEP_SIZE mode
This implementation relies on the earlier implementation
of FALLOC_FL_ZERO_RANGE. It stores zeros in the specified range
and leaves the file with the same content as when executed
under EXT4 filesystem.

This functionality comes with the following restrictions:
* After calling fallocate with this mode, llseek will not be
functioning as expected when SEEK_DATA or SEEK_HOLE are specified
as whence option.
* Tests checking number of blocks allocated via stat() system call
will be failing. UBIFS already struggles with reporting this and
in addition punched holes won't be taken into account.

Listed restrictions do not affect current state of Android.

Change-Id: I7e439edc7222b4d4ea040d498eeb5ff1acc64acd
Signed-off-by: Dragan Cecavac <dragan.cecavac@imgtec.com>