sound: soc: jz4780: npm801: Enable speakers audio and headphones detection
NPM801 uses the same output route for speakers and headphones.
Selection between those is performed by GPIO configuration,
where it's actually possible to have both enabled at the same time.

In comparison to CI20, where a GPIO is dedicated for
headphones detection, on NPM801 it has to be done by checking
the value of the status register.
This value check has been done in a thread which is active
when headphones or speakers stream is enabled and it will select
the appropriate output.

Change-Id: I9a356347d81f83ee118e9fa60e26a549462e9d13
Signed-off-by: Dragan Cecavac <dragan.cecavac@imgtec.com>