dma-jz4780: Fix calculating transfer residue
Granularity of the reported transfer residue for jz4780 DMA controller is
DMA_RESIDUE_GRANULARITY_BURST, which means residue is updated after each
transferred burst. Residue is calculated based on number of blocks to
transfer for all remaining descriptors in current transfer, and value of
DTC register of DMA channel which stores number of block that were
successfully transferred in previous burst.

Residue was not being calculated correctly because of few mistakes, which
this patch corrects. Number of blocks to tranfer in one descriptor is
stored in the lowest 24 bits of dtc field, so upper 8 bits that represent
offset of next descriptor shouldn't be considered. Also, it was incorrectly
assumed that DTC register of DMA channel stores number of blocks that
were not successfully transferred in previous burst, while it actually
stores number of blocks successfully transferred. Finally, incorrect
next descriptor to transfer was being passed to function that calculates
residue.

This patch also sets calculated residue by using dma_set_residue() API
instead of directly accessing residue field of structure dma_tx_state.

Change-Id: I5b4d4e1916a1b9b1020c110c5811da427efa3289
Signed-off-by: Darko Ristivojevic <darko.ristivojevic@mips.com>